ommy_poland/requirements.txt
Django==4.0.1
psycopg2==2.9.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
